Fix compilation under Sun Studio 12 U1 on OpenSolaris
Thanks to tropikhajma for testing and patch

BUG: 213492

svn path=/trunk/KDE/kdeedu/kstars/; revision=1118376

+++ kstars/skymap.cpp
@@ -1304,7 +1304,7 @@ void SkyMap::forceUpdate( bool now )
 }
 
 float SkyMap::fov() {
-    float diagonalPixels = sqrt( width() * width() + height() * height() );
+    float diagonalPixels = sqrt(static_cast<double>( width() * width() + height() * height() ));
     return diagonalPixels / ( 2 * Options::zoomFactor() * dms::DegToRad );
 }
